order frá checkout saveast í database og eytt öllu úr cart

+++ BookCave/Models/ViewModels/OrderDetailViewModel.cs
@@ -7,7 +7,7 @@ namespace BookCave.Models.ViewModels
     public class OrderDetailViewModel
     {
         public int Id { get; set; }
-        public int Total { get; set; }
+        public double Total { get; set; }
         public DateTime PaidDate { get; set; }
         public string CustomerId { get; set; }
         public List<OrderItem> OrderItems { get; set; }
